Move buildnum.c to src/base. Do not take local time into account when building version number - it was weird (quake uses 4 digits for that, we could easily have 7-8). Fix dynamic lights. QW client printed version Twilight version as 0 - fixed. Move COM_BlockSequenceCRCByte to crc.c

+++ src/nq/sys.c
@@ -362,7 +362,7 @@ Sys_DoubleTime (void)
 
 	last = now;
 
-	return epoch + now / 1000.0;
+	return epoch + (double)(now / 1000.0);
 }
 
 char *
